update[proxy]: make lint
